tari-project/wallet-android#487
Add bottom bar navigation at the home page
