+++ app/src/main/java/com/tari/android/wallet/di/ApplicationComponent.kt
@@ -43,7 +43,6 @@ import com.tari.android.wallet.ui.activity.SplashActivity
 import com.tari.android.wallet.ui.activity.debug.DebugActivity
 import com.tari.android.wallet.ui.activity.home.HomeActivity
 import com.tari.android.wallet.ui.activity.onboarding.OnboardingFlowActivity
-import com.tari.android.wallet.ui.activity.profile.WalletInfoActivity
 import com.tari.android.wallet.ui.activity.qr.QRScannerActivity
 import com.tari.android.wallet.ui.activity.restore.WalletRestoreActivity
 import com.tari.android.wallet.ui.activity.send.SendTariActivity
@@ -53,6 +52,7 @@ import com.tari.android.wallet.ui.fragment.debug.DebugLogFragment
 import com.tari.android.wallet.ui.fragment.onboarding.CreateWalletFragment
 import com.tari.android.wallet.ui.fragment.onboarding.IntroductionFragment
 import com.tari.android.wallet.ui.fragment.onboarding.LocalAuthFragment
+import com.tari.android.wallet.ui.fragment.profile.WalletInfoFragment
 import com.tari.android.wallet.ui.fragment.restore.ChooseRestoreOptionFragment
 import com.tari.android.wallet.ui.fragment.restore.EnterRestorationPasswordFragment
 import com.tari.android.wallet.ui.fragment.send.AddAmountFragment
@@ -60,9 +60,10 @@ import com.tari.android.wallet.ui.fragment.send.AddNoteFragment
 import com.tari.android.wallet.ui.fragment.send.AddRecipientFragment
 import com.tari.android.wallet.ui.fragment.send.FinalizeSendTxFragment
 import com.tari.android.wallet.ui.fragment.settings.AllSettingsFragment
+import com.tari.android.wallet.ui.fragment.settings.backup.BackupSettingsFragment
 import com.tari.android.wallet.ui.fragment.settings.backup.ChangeSecurePasswordFragment
 import com.tari.android.wallet.ui.fragment.settings.backup.EnterCurrentPasswordFragment
-import com.tari.android.wallet.ui.fragment.settings.backup.BackupSettingsFragment
+import com.tari.android.wallet.ui.fragment.tx.TxListFragment
 import com.tari.android.wallet.util.SharedPrefsWrapper
 import dagger.Component
 import java.io.File
@@ -110,7 +111,6 @@ internal interface ApplicationComponent {
     fun inject(activity: HomeActivity)
     fun inject(activity: QRScannerActivity)
     fun inject(activity: SendTariActivity)
-    fun inject(activity: WalletInfoActivity)
     fun inject(activity: TxDetailActivity)
     fun inject(activity: DebugActivity)
 
@@ -127,6 +127,8 @@ internal interface ApplicationComponent {
     fun inject(fragment: DebugLogFragment)
     fun inject(fragment: BaseNodeConfigFragment)
     fun inject(fragment: AllSettingsFragment)
+    fun inject(fragment: WalletInfoFragment)
+    fun inject(fragment: TxListFragment)
     /**
      * Backup.
      */
