+++ app/src/main/java/com/tari/android/wallet/ui/fragment/store/StoreFragment.kt
@@ -33,52 +33,43 @@
 package com.tari.android.wallet.ui.fragment.store
 
 import android.annotation.SuppressLint
-import android.app.Dialog
 import android.content.Intent
-import android.graphics.Color
 import android.os.Bundle
-import android.view.KeyEvent.KEYCODE_BACK
 import android.view.LayoutInflater
-import android.view.MotionEvent
 import android.view.View
 import android.view.ViewGroup
-import android.view.ViewGroup.LayoutParams.MATCH_PARENT
-import android.widget.FrameLayout
 import android.widget.Toast
 import androidx.appcompat.view.ContextThemeWrapper
-import androidx.coordinatorlayout.widget.CoordinatorLayout
-import androidx.core.widget.NestedScrollView
-import androidx.fragment.app.DialogFragment
-import com.google.android.material.bottomsheet.BottomSheetBehavior.*
-import com.google.android.material.bottomsheet.BottomSheetDialog
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
+import androidx.fragment.app.Fragment
 import com.tari.android.wallet.R
 import com.tari.android.wallet.R.color.purple
 import com.tari.android.wallet.R.drawable.*
 import com.tari.android.wallet.R.string.store_no_application_to_open_the_link_error
 import com.tari.android.wallet.R.string.ttl_store_url
-import com.tari.android.wallet.databinding.DialogStoreBinding
+import com.tari.android.wallet.databinding.FragmentStoreBinding
 import com.tari.android.wallet.ui.extension.color
 import com.tari.android.wallet.ui.extension.drawable
 import com.tari.android.wallet.ui.extension.gone
 import com.tari.android.wallet.ui.extension.string
 import com.tari.android.wallet.ui.util.UiUtil
 
-class StoreDialogFragment private constructor() : BottomSheetDialogFragment() {
+class StoreFragment @Deprecated(
+    """Use newInstance() and supply all the 
+necessary data via arguments instead, as fragment's default no-op constructor is used by the 
+framework for UI tree rebuild on configuration changes"""
+) constructor() : Fragment() {
 
-    private var _ui: DialogStoreBinding? = null
-    private val ui get() = _ui!!
-    private lateinit var behavior: LockBottomSheetBehavior<*>
+    private lateinit var ui: FragmentStoreBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? = DialogStoreBinding.inflate(
+    ): View? = FragmentStoreBinding.inflate(
         LayoutInflater.from(ContextThemeWrapper(requireContext(), R.style.AppTheme)),
         container,
         false
-    ).also { _ui = it }.root
+    ).also { ui = it }.root
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
@@ -94,7 +85,6 @@ class StoreDialogFragment private constructor() : BottomSheetDialogFragment() {
         ui.browserForwardCtaView.setOnClickListener {
             ui.webView.apply { if (canGoForward()) goForward() }
         }
-        ui.closeCtaView.setOnClickListener { behavior.state = STATE_HIDDEN }
         ui.shareCtaView.setOnClickListener { shareStoreLink() }
     }
 
@@ -145,74 +135,14 @@ class StoreDialogFragment private constructor() : BottomSheetDialogFragment() {
         )
     }
 
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog =
-        BottomSheetDialog(requireContext(), R.style.StoreBottomSlideDialog)
-            .apply { setOnShowListener { setupDialog(this) } }
-
-    private fun setupDialog(bottomSheetDialog: BottomSheetDialog) {
-        configureScrollBehavior(bottomSheetDialog.findViewById(R.id.design_bottom_sheet)!!)
-        bottomSheetDialog.setOnKeyListener { _, keyCode, _ ->
-            (keyCode == KEYCODE_BACK && behavior.state != STATE_HIDDEN &&
-                    behavior.state != STATE_COLLAPSED).also {
-                if (it) behavior.state = STATE_HIDDEN
-            }
-        }
-    }
-
-    private fun configureScrollBehavior(bottomSheet: FrameLayout) {
-        behavior = LockBottomSheetBehavior<View>()
-        behavior.isHideable = true
-        behavior.skipCollapsed = true
-        behavior.state = STATE_EXPANDED
-        val layoutParams = bottomSheet.layoutParams as CoordinatorLayout.LayoutParams
-        layoutParams.height = MATCH_PARENT
-        layoutParams.behavior = behavior
-        bottomSheet.layoutParams = layoutParams
-        disallowDismissOnScrollViewInteraction(ui.scrollView, behavior)
-        disallowDismissOnScrollViewInteraction(ui.controlsScrollView, behavior)
-        val rootView = bottomSheet.parent as View
-        rootView.setBackgroundColor(Color.BLACK)
-        behavior.addBottomSheetCallback(DismissOnHide(this, rootView))
-    }
-
-    private fun disallowDismissOnScrollViewInteraction(
-        view: NestedScrollView,
-        behavior: LockBottomSheetBehavior<*>
-    ) {
-        view.apply {
-            setOnTouchListener { _, event ->
-                behavior.allowUserDragging = false
-                if (event.action and MotionEvent.ACTION_MASK == MotionEvent.ACTION_UP) {
-                    behavior.allowUserDragging = true
-                }
-                false
-            }
-        }
-    }
-
     override fun onDestroyView() {
-        super.onDestroyView()
         ui.webView.destroy()
-        _ui = null
-    }
-
-    private class DismissOnHide(private val fragment: DialogFragment, private val rootView: View) :
-        BottomSheetCallback() {
-        override fun onStateChanged(bottomSheet: View, newState: Int) {
-            if (newState == STATE_HIDDEN || newState == STATE_COLLAPSED) {
-                fragment.dismiss()
-            }
-        }
-
-        override fun onSlide(bottomSheet: View, slideOffset: Float) {
-            val color: Int = Color.argb((slideOffset.coerceIn(0F, 1F) * 255).toInt(), 0, 0, 0)
-            rootView.setBackgroundColor(color)
-        }
-
+        super.onDestroyView()
     }
 
     companion object {
-        fun newInstance() = StoreDialogFragment()
+        @Suppress("DEPRECATION")
+        fun newInstance() = StoreFragment()
     }
 
 }
