+++ app/src/main/res/color/home_bottom_nav_view_icon_selector.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+    <item android:state_checked="true" android:color="@color/black" />
+    <item android:state_checked="false" android:color="@color/black"/>
+</selector>
