+++ app/src/main/res/drawable/icon_home.xml
@@ -1,0 +1,14 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="28dp"
+    android:height="23dp"
+    android:viewportWidth="28"
+    android:viewportHeight="23">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M22.77,9.615c-0.035,-0.028 -0.08,-0.043 -0.123,-0.043 -0.03,0 -0.06,0.007 -0.087,0.021 -0.07,0.033 -0.113,0.103 -0.113,0.18v10.971L18.12,20.744L18.12,14.869c0,-0.482 -0.393,-0.876 -0.875,-0.876h-6.077c-0.482,0 -0.875,0.394 -0.875,0.876v5.876L5.968,20.745L5.968,9.772c0,-0.076 -0.044,-0.146 -0.113,-0.179 -0.027,-0.014 -0.057,-0.02 -0.087,-0.02 -0.044,0 -0.088,0.014 -0.124,0.042l-1.351,1.065c-0.047,0.04 -0.076,0.096 -0.076,0.158L4.217,21.619c0,0.483 0.393,0.876 0.876,0.876h6.076c0.482,0 0.875,-0.393 0.876,-0.876v-5.876h4.326v5.876c0,0.483 0.392,0.876 0.875,0.876h6.076c0.483,0 0.876,-0.393 0.876,-0.876L24.198,10.839c0,-0.063 -0.03,-0.12 -0.076,-0.159L22.77,9.615z"
+      android:fillType="evenOdd"/>
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M27.581,10.797L14.74,0.659c-0.17,-0.112 -0.352,-0.17 -0.556,-0.17 -0.178,0.004 -0.35,0.062 -0.522,0.18L0.831,10.798c-0.375,0.303 -0.442,0.857 -0.148,1.238 0.156,0.202 0.416,0.323 0.695,0.323 0.203,0 0.398,-0.065 0.547,-0.182l12.282,-9.693 12.282,9.693c0.15,0.117 0.341,0.182 0.541,0.182 0.274,0 0.531,-0.121 0.69,-0.321 0.147,-0.187 0.213,-0.436 0.18,-0.684 -0.032,-0.232 -0.147,-0.431 -0.319,-0.557"
+      android:fillType="evenOdd"/>
+</vector>
