+++ app/src/main/res/drawable/icon_profile.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="21dp"
+    android:height="23dp"
+    android:viewportWidth="21"
+    android:viewportHeight="23">
+  <path
+      android:fillColor="#FF000000"
+      android:pathData="M2.047,20.64c0.569,-3.364 4.035,-4.887 8.225,-4.887 4.19,0 7.656,1.523 8.224,4.887L2.047,20.64zM1.73,22.107L0.5,22.107c0,-5.398 4.375,-7.819 9.772,-7.819s9.772,2.421 9.772,7.82L1.73,22.108zM10.272,10.868c2.429,0 4.398,-1.969 4.398,-4.397 0,-2.429 -1.97,-4.398 -4.398,-4.398 -2.43,0 -4.397,1.969 -4.397,4.398 0,2.428 1.968,4.397 4.397,4.397m0,1.466c3.238,0 5.864,-2.625 5.864,-5.863C16.136,3.232 13.51,0.607 10.272,0.607S4.409,3.232 4.409,6.471c0,3.238 2.625,5.863 5.863,5.863"
+      android:fillType="evenOdd"/>
+</vector>
