+++ app/src/main/res/drawable/icon_settings.xml
@@ -1,0 +1,18 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <group>
+    <clip-path
+        android:pathData="M0.899,0.8L23.878,0.8 23.878,23.78 0.899,23.78z"/>
+    <path
+        android:pathData="M19.605,14.608c-0.125,0.391 -0.285,0.777 -0.477,1.147 -0.29,0.566 -0.183,1.247 0.267,1.695l0.96,1.025c-0.515,0.665 -1.116,1.265 -1.85,1.78l-0.956,-0.959c-0.27,-0.269 -0.648,-0.423 -1.035,-0.423 -0.23,0 -0.458,0.054 -0.66,0.158 -0.366,0.188 -0.753,0.348 -1.148,0.475 -0.605,0.194 -1.012,0.75 -1.012,1.386l-0.046,1.405c-0.413,0.052 -0.838,0.08 -1.26,0.08 -0.421,0 -0.845,-0.028 -1.304,-0.13v-1.355c0,-0.636 -0.407,-1.192 -1.012,-1.386 -0.39,-0.125 -0.777,-0.285 -1.148,-0.475 -0.207,-0.107 -0.438,-0.163 -0.67,-0.163 -0.385,0 -0.75,0.152 -1.025,0.427l-1.027,0.962c-0.665,-0.516 -1.264,-1.116 -1.776,-1.849l0.957,-0.958c0.449,-0.449 0.554,-1.13 0.264,-1.696 -0.189,-0.367 -0.349,-0.752 -0.474,-1.148 -0.194,-0.604 -0.752,-1.01 -1.386,-1.01L2.38,13.55c-0.105,-0.836 -0.105,-1.684 0.052,-2.565h1.354c0.635,0 1.192,-0.407 1.385,-1.013 0.127,-0.394 0.287,-0.78 0.476,-1.147 0.29,-0.565 0.184,-1.246 -0.265,-1.696L4.42,6.104c0.518,-0.666 1.117,-1.266 1.85,-1.777l0.959,0.958c0.268,0.268 0.646,0.423 1.034,0.423 0.23,0 0.457,-0.055 0.66,-0.159 0.367,-0.188 0.752,-0.348 1.149,-0.475 0.605,-0.194 1.012,-0.75 1.012,-1.386l0.043,-1.406c0.415,-0.05 0.84,-0.077 1.261,-0.077 0.422,0 0.847,0.026 1.306,0.13v1.353c0,0.636 0.406,1.192 1.012,1.386 0.395,0.126 0.78,0.286 1.149,0.475 0.2,0.104 0.43,0.16 0.66,0.16 0.386,0 0.764,-0.156 1.034,-0.425l1.026,-0.96c0.663,0.515 1.264,1.115 1.777,1.848l-0.957,0.957c-0.45,0.451 -0.555,1.131 -0.265,1.695 0.188,0.367 0.349,0.754 0.475,1.148 0.194,0.606 0.75,1.013 1.386,1.013l1.405,0.044c0.104,0.837 0.104,1.685 -0.052,2.566L20.99,13.595c-0.635,0 -1.191,0.408 -1.385,1.013m4.183,-3.752c-0.09,-0.727 -0.712,-1.275 -1.444,-1.275l-1.404,-0.035c-0.149,-0.469 -0.338,-0.927 -0.553,-1.423l0.956,-0.958c0.52,-0.518 0.572,-1.344 0.12,-1.923 -0.588,-0.757 -1.27,-1.44 -2.027,-2.029 -0.256,-0.198 -0.573,-0.307 -0.893,-0.307 -0.39,0 -0.755,0.152 -1.03,0.428l-1.018,0.966c-0.436,-0.224 -0.895,-0.414 -1.397,-0.613L15.098,2.334c0,-0.733 -0.549,-1.353 -1.276,-1.444 -0.47,-0.059 -0.952,-0.09 -1.433,-0.09 -0.48,0 -0.963,0.031 -1.435,0.09C10.226,0.981 9.68,1.601 9.68,2.334l-0.037,1.403c-0.47,0.15 -0.928,0.34 -1.423,0.555l-0.957,-0.958c-0.276,-0.276 -0.64,-0.427 -1.029,-0.427 -0.32,0 -0.638,0.11 -0.894,0.306 -0.756,0.59 -1.44,1.273 -2.027,2.03 -0.45,0.577 -0.399,1.404 0.12,1.923L4.4,8.182c-0.225,0.437 -0.415,0.895 -0.613,1.4L2.433,9.582c-0.733,0 -1.354,0.547 -1.445,1.274 -0.12,0.951 -0.12,1.916 0,2.87 0.092,0.725 0.712,1.273 1.445,1.273l1.403,0.037c0.15,0.468 0.34,0.926 0.554,1.422l-0.957,0.958c-0.517,0.517 -0.57,1.343 -0.12,1.921 0.588,0.757 1.27,1.44 2.027,2.028 0.256,0.2 0.572,0.308 0.894,0.308 0.388,0 0.753,-0.15 1.03,-0.426l1.018,-0.968c0.435,0.225 0.893,0.414 1.398,0.613v1.355c0,0.731 0.547,1.351 1.274,1.443 0.471,0.06 0.954,0.09 1.434,0.09s0.963,-0.03 1.435,-0.09c0.726,-0.092 1.274,-0.713 1.274,-1.444l0.037,-1.404c0.47,-0.15 0.928,-0.34 1.422,-0.553l0.957,0.957c0.275,0.275 0.64,0.427 1.03,0.427 0.32,0 0.637,-0.109 0.893,-0.307 0.758,-0.59 1.44,-1.273 2.029,-2.029 0.45,-0.579 0.398,-1.405 -0.12,-1.922l-0.967,-1.018c0.224,-0.435 0.413,-0.893 0.613,-1.398h1.354c0.733,0 1.353,-0.548 1.444,-1.275 0.118,-0.95 0.118,-1.917 0,-2.868"
+        android:fillColor="#000"
+        android:fillType="evenOdd"/>
+  </group>
+  <path
+      android:pathData="M12.388,9.228c0.817,0 1.586,0.318 2.165,0.897 0.578,0.578 0.897,1.346 0.897,2.165 0,0.817 -0.319,1.586 -0.897,2.165 -0.578,0.578 -1.347,0.896 -2.165,0.896 -0.818,0 -1.586,-0.318 -2.165,-0.896 -0.578,-0.579 -0.896,-1.348 -0.896,-2.165 0,-0.819 0.318,-1.587 0.896,-2.165 0.58,-0.579 1.348,-0.897 2.165,-0.897m0,-1.403c-1.193,0 -2.314,0.465 -3.157,1.307 -0.843,0.844 -1.308,1.965 -1.308,3.158 0,1.193 0.465,2.314 1.308,3.158 0.844,0.842 1.966,1.307 3.157,1.307 1.192,0 2.314,-0.465 3.158,-1.307 0.843,-0.844 1.308,-1.965 1.308,-3.158 0,-1.193 -0.465,-2.314 -1.308,-3.158 -0.844,-0.842 -1.965,-1.307 -3.158,-1.307"
+      android:fillColor="#000"
+      android:fillType="evenOdd"/>
+</vector>
