+++ app/src/main/res/layout/activity_settings.xml
@@ -5,5 +5,5 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/black"
-    tools:context=".ui.activity.settings.SettingsActivity"
+    tools:context=".ui.activity.settings.BackupSettingsActivity"
     tools:viewBindingIgnore="true" />
