+++ app/src/main/res/values/colors.xml
@@ -53,6 +53,7 @@
     <color name="home_tx_section_header_title">#B2B2B2</color>
     <color name="home_black_overlay">#FF000000</color>
     <color name="create_emoji_id_bg">#f6f6f6</color>
+    <color name="home_selected_nav_item">#9330FF</color>
 
     <!-- send tari - add recipient -->
     <color name="add_recipient_hint">#B2B2B2</color>
