+++ app/src/main/res/values/styles.xml
@@ -29,19 +29,9 @@
         <item name="android:windowAnimationStyle">@style/BottomSlideDialogAnimation</item>
     </style>
 
-    <style name="StoreBottomSlideDialog" parent="android:Theme.Panel">
-        <item name="android:windowAnimationStyle">@style/StoreBottomSlideDialogAnimation</item>
-        <item name="android:windowTranslucentNavigation">true</item>
-    </style>
-
     <style name="BottomSlideDialogAnimation">
         <item name="android:windowEnterAnimation">@anim/dialog_slide_from_bottom</item>
         <item name="android:windowExitAnimation">@anim/dialog_slide_to_bottom</item>
-    </style>
-
-    <style name="StoreBottomSlideDialogAnimation">
-        <item name="android:windowEnterAnimation">@anim/dialog_slide_from_bottom</item>
-        <item name="android:windowExitAnimation">@null</item>
     </style>
 
     <style name="PurpleCheckbox" parent="AppTheme">
